+++ product/dashboard/src/test/java/io/codekvast/dashboard/webapp/impl/WebappServiceImplTest.java
@@ -5,7 +5,8 @@ import io.codekvast.dashboard.customer.CustomerService;
 import io.codekvast.dashboard.customer.PricePlan;
 import io.codekvast.dashboard.customer.PricePlanDefaults;
 import io.codekvast.dashboard.security.CustomerIdProvider;
-import io.codekvast.dashboard.webapp.WebappService;
+import io.codekvast.dashboard.util.TimeService;
+import io.codekvast.dashboard.webapp.model.methods.GetMethodsRequest;
 import io.codekvast.dashboard.webapp.model.status.GetStatusResponse1;
 import org.junit.Before;
 import org.junit.Test;
@@ -38,21 +39,27 @@ public class WebappServiceImplTest {
     @Mock
     private CustomerIdProvider customerIdProvider;
 
-    private WebappService webappService;
+    @Mock
+    private TimeService timeService;
+
+    private Instant now = Instant.now();
+
+    private WebappServiceImpl webappService;
 
     @Before
-    public void beforeTest() throws Exception {
+    public void beforeTest() {
         MockitoAnnotations.initMocks(this);
-        webappService = new WebappServiceImpl(jdbcTemplate, customerIdProvider, customerService);
+        webappService = new WebappServiceImpl(jdbcTemplate, customerIdProvider, customerService, timeService);
+        when(timeService.now()).thenReturn(now);
+        when(timeService.currentTimeMillis()).thenReturn(now.toEpochMilli());
     }
 
     @Test
-    public void should_getStatus_inside_trial_period() throws Exception {
+    public void should_getStatus_inside_trial_period() {
         // given
         when(customerIdProvider.getCustomerId()).thenReturn(1L);
 
         PricePlanDefaults ppd = PricePlanDefaults.TEST;
-        Instant now = Instant.now();
         Instant collectionStartedAt = now.minus(3, DAYS);
         Instant trialPeriodEndsAt = now.plus(3, DAYS);
         CustomerData customerData = CustomerData.builder()
@@ -66,7 +73,7 @@ public class WebappServiceImplTest {
         when(customerService.getCustomerDataByCustomerId(eq(1L))).thenReturn(customerData);
         when(customerService.countMethods(eq(1L))).thenReturn(1000);
 
-            // when
+        // when
         GetStatusResponse1 status = webappService.getStatus();
 
         // then
@@ -91,12 +98,11 @@ public class WebappServiceImplTest {
     }
 
     @Test
-    public void should_getStatus_after_trial_period() throws Exception {
+    public void should_getStatus_after_trial_period() {
         // given
         when(customerIdProvider.getCustomerId()).thenReturn(1L);
 
         PricePlanDefaults ppd = PricePlanDefaults.TEST;
-        Instant now = Instant.now();
         Instant collectionStartedAt = now.minus(3, DAYS);
         Instant trialPeriodEndsAt = now.minus(1, MILLIS);
         CustomerData customerData = CustomerData.builder()
@@ -122,7 +128,7 @@ public class WebappServiceImplTest {
     }
 
     @Test
-    public void should_getStatus_not_started_no_trial_period() throws Exception {
+    public void should_getStatus_not_started_no_trial_period() {
         // given
         when(customerIdProvider.getCustomerId()).thenReturn(1L);
 
@@ -147,5 +153,69 @@ public class WebappServiceImplTest {
         assertThat(status.getTrialPeriodEndsAtMillis(), is(nullValue()));
         assertThat(status.getTrialPeriodPercent(), is(nullValue()));
         assertThat(status.getTrialPeriodExpired(), is(false));
+    }
+
+    @Test
+    public void should_build_default_select_statement() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults());
+        assertThat(whereClause, is("m.signature LIKE '%'"));
+    }
+
+    @Test
+    public void should_build_normalized_where_clause_with_redundant_wildcards() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults().toBuilder()
+                                                                             .signature("%%").build());
+        assertThat(whereClause, is("m.signature LIKE '%'"));
+    }
+
+    @Test
+    public void should_build_normalized_where_clause_with_explicit_signature() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults().toBuilder()
+                                                                             .normalizeSignature(true)
+                                                                             .signature("foobar").build());
+        assertThat(whereClause, is("m.signature LIKE '%foobar%'"));
+    }
+
+    @Test
+    public void should_build_where_clause_with_unnormalized_signature() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults().toBuilder()
+                                                                             .normalizeSignature(false)
+                                                                             .signature("foobar").build());
+        assertThat(whereClause, is("m.signature = 'foobar'"));
+    }
+
+    @Test
+    public void should_build_where_clause_for_only_invoked_before() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults().toBuilder()
+                                                                             .onlyInvokedBeforeMillis(100_000_000L)
+                                                                             .build());
+        assertThat(whereClause, is("m.signature LIKE '%' AND i.invokedAtMillis <= 100000000"));
+    }
+
+    @Test
+    public void should_build_where_clause_for_only_invoked_after() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults().toBuilder()
+                                                                             .onlyInvokedAfterMillis(100_000_000L)
+                                                                             .build());
+        assertThat(whereClause, is("m.signature LIKE '%' AND i.invokedAtMillis >= 100000000"));
+    }
+
+    @Test
+    public void should_build_where_clause_for_only_invoked_between() {
+        String whereClause = webappService.buildWhereClause(GetMethodsRequest.defaults().toBuilder()
+                                                                             .onlyInvokedAfterMillis(100_000_000L)
+                                                                             .onlyInvokedBeforeMillis(200_000_000L)
+                                                                             .build());
+        assertThat(whereClause, is("m.signature LIKE '%' AND (i.invokedAtMillis BETWEEN 100000000 AND 200000000)"));
+    }
+
+    @Test
+    public void should_detect_synthetic_method_containing_dot_dot() {
+        assertThat(webappService.isSyntheticMethod("foo..bar"), is(true));
+    }
+
+    @Test
+    public void should_not_detect_synthetic_method_not_containing_dot_dot() {
+        assertThat(webappService.isSyntheticMethod("foo.bar"), is(false));
     }
 }